backend-thinkphp/scripts/requirements.txt
paho-mqtt>=2.0.0
requests>=2.28.0
